Split the SU2 ASCII finite volume reader in a common base class and a class that is specific to the reading of the SU2 ASCII grid for the finite volume solver
