+++ Common/src/geometry/meshreader/meson.build
@@ -3,4 +3,5 @@ common_src += files(['CBoxMeshReaderFVM.cpp',
                      'CCGNSMeshReaderFVM.cpp',
                      'CMeshReaderBase.cpp',
                      'CRectangularMeshReaderFVM.cpp',
+                     'CSU2ASCIIMeshReaderBase.cpp',
                      'CSU2ASCIIMeshReaderFVM.cpp'])
